modifies DMXConnection to allow error handling to propogate up out of module.
Specifically adds handling of serial connection during instatiation of DMXConnection. This error comes from the `serial` library and is emitted when the serial connection fails.

Adds a new member variable to DMXConnection, softfail, which allows the DMXConnection object to NOT exit the python application BUT also retains the original interface (`softfail` defaults to `False`).