Add hard system reboot for failing Serial communication
IMPORTANT: This is a hack
Adds a try/catch to com.write() in render() and a write timeout to serial connection. Now, if the com.write() times out (and a flag is set) the system is forced to reboot.

Note, this is all because the thread which runs the ramp() method cannot raise exceptions to the top level app.
The enttec device has a serial command which allows for a smooth transition between scenes, this would allow the removal of the thread and thus proper exception handling at the app level.